+++ apps/yjshd/app/(dynamic)/3.%EC%BD%A9%ED%8C%A5%EC%A7%88%ED%99%98%20%EA%B0%95%EC%A2%8C/3.%ED%98%88%EC%95%A1%ED%88%AC%EC%84%9D%201,2,3/VideoSection.tsx
@@ -1,36 +1,14 @@
 // app/playlist/VideosSection.tsx
-import PaginationLink from "@/components/PaginationLink";
+import Pagination from "@/components/Pagination";
 import YoutubePlaylist from "@/components/YoutubePlaylist";
 import { getAllVideos } from "@/lib/youtube";
 import { Badge } from "@workspace/ui/components/badge";
-import { Pagination, PaginationContent, PaginationItem } from "@workspace/ui/components/pagination";
-import { ChevronLeft, ChevronRight } from "lucide-react";
 
 interface VideosSectionProps {
   page?: string;
 }
 
 const pageSize = 12; // 한 페이지당 표시할 비디오 수
-
-/* ---------- Pagination 계산 ---------- */
-function buildPages(total: number, current: number) {
-  const out: (number | "…")[] = [];
-  const sibling = 1;
-  const first = 1;
-  const last = total;
-
-  out.push(first);
-
-  const left = Math.max(current - sibling, first + 1);
-  const right = Math.min(current + sibling, last - 1);
-
-  if (left > first + 1) out.push("…");
-  for (let i = left; i <= right; i++) out.push(i);
-  if (right < last - 1) out.push("…");
-
-  if (last !== first) out.push(last);
-  return out;
-}
 
 export default async function VideosSection({ page }: VideosSectionProps) {
   const currentPage = Math.max(parseInt(page || "1", 10), 1);
@@ -43,11 +21,6 @@ export default async function VideosSection({ page }: VideosSectionProps) {
   const startIndex = (currentPage - 1) * pageSize;
   const endIndex = startIndex + pageSize;
   const currentPageVideos = allVideos.slice(startIndex, endIndex);
-
-  const pages = buildPages(totalPages, currentPage);
-
-  const prevHref = currentPage > 1 ? `?page=${currentPage - 1}` : undefined;
-  const nextHref = currentPage < totalPages ? `?page=${currentPage + 1}` : undefined;
 
   return (
     <div className="space-y-6 relative">
@@ -70,54 +43,7 @@ export default async function VideosSection({ page }: VideosSectionProps) {
       </article>
 
       {/* 페이지네이션 컨트롤 */}
-      <nav aria-label="비디오 페이지" className="mt-8">
-        <Pagination>
-          <PaginationContent className="flex items-center justify-center gap-1">
-            {/* First */}
-            <PaginationItem className="hidden sm:block">
-              <PaginationLink href="?page=1" aria-label="첫 페이지" className={`${prevHref ? "" : "pointer-events-none opacity-50"}`}>
-                «
-              </PaginationLink>
-            </PaginationItem>
-
-            {/* Prev */}
-            <PaginationItem className="hidden sm:block">
-              <PaginationLink href={prevHref || "#"} aria-disabled={!prevHref} aria-label="이전 페이지" className={`${prevHref ? "" : "pointer-events-none opacity-50"}`}>
-                <ChevronLeft className="h-4 w-4" />
-              </PaginationLink>
-            </PaginationItem>
-
-            {/* 숫자 & Ellipsis */}
-            {pages.map((p, idx) =>
-              typeof p === "number" ? (
-                <PaginationItem key={idx}>
-                  <PaginationLink href={`?page=${p}`} isActive={p === currentPage}>
-                    {p}
-                  </PaginationLink>
-                </PaginationItem>
-              ) : (
-                <li key={idx} className="px-3 py-2 min-w-11 min-h-11 flex items-center justify-center text-gray-400 select-none">
-                  …
-                </li>
-              )
-            )}
-
-            {/* Next */}
-            <PaginationItem className="hidden sm:block">
-              <PaginationLink href={nextHref || "#"} aria-disabled={!nextHref} aria-label="다음 페이지" className={`${nextHref ? "" : "pointer-events-none opacity-50"}`}>
-                <ChevronRight className="h-4 w-4" />
-              </PaginationLink>
-            </PaginationItem>
-
-            {/* Last */}
-            <PaginationItem className="hidden sm:block">
-              <PaginationLink href={`?page=${totalPages}`} aria-label="마지막 페이지" className={`${nextHref ? "" : "pointer-events-none opacity-50"}`}>
-                »
-              </PaginationLink>
-            </PaginationItem>
-          </PaginationContent>
-        </Pagination>
-      </nav>
+      <Pagination currentPage={currentPage} totalPages={totalPages} baseUrl="/3.콩팥질환 강좌/3.혈액투석 1,2,3" className="mt-8" />
     </div>
   );
 }
